Adjust "Compile Examples" workflow for Seeed boards
The "Compile Examples" GitHub Actions workflow compiles example sketches of the platform bundled libraries for each of the platform's boards on every push, pull request, and periodically. This provides a basic "smoke test" for the platform.

The workflow originated from the "Arduino Mbed OS Boards" platform, and thus was configured to compile for those boards. It must be reconfigured here to install the correct toolchain and compile for Seeed Studio's boards of the platform.